Rework some more retworkx usage to fix some issues
This commit continues on the work to adapt the dag circuit to use
retworkx PyDAG instead of networkx. It's still not fully working, but it
fixes the common case and about 100+ of the ~400 failing unit tests are
fixed by this. It did also find several bugs in retworkx which will be
fixed shortly in a pending release.